pyproject.toml
[tool.poetry]
name = "eco-update-synthesizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
bs4 = "^0.0.1"
justext = "^2.2.0"
gensim = "^4.0.1"
geotext = "^0.4.0"
newspaper3k = "^0.2.8"
requests-html = "^0.10.0"
geocoder = "^1.38.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
